# next-rms-app/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SmartRent API - Tài liệu API cho Mobile App",
    "version": "1.0.0",
    "description": "Tài liệu mô tả toàn bộ các API mà Mobile App (cho Manager và Tenant) cần kết nối. Tất cả các API cần xác thực đều phải gửi kèm `Authorization: Bearer <access_token>` trong header.\n\n## Cách lấy Access Token\nGọi API `POST /api/auth/login` với số điện thoại và mật khẩu. Token sẽ được trả về trong trường `access_token`.\n\n## Phân quyền (Roles)\n- **super_admin**: Toàn quyền.\n- **manager**: Chỉ xem/thao tác dữ liệu thuộc chi nhánh được gán.\n- **tenant**: Chỉ xem/thao tác dữ liệu của phòng mình đang thuê."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Môi trường phát triển (Local)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Nhập Access Token (lấy từ API /api/auth/login)"
      }
    },
...
  },
  "security": [],
  "paths": {
...
    "/api/users/create": {
      "post": {
        "tags": ["👤 Quản lý Người dùng (Users)"],
        "summary": "Tạo tài khoản mới (Manager/Tenant)",
        "description": "Tạo tài khoản mới mà không cần xác thực OTP SMS. **Phân quyền:**\n- **super_admin** có thể tạo tài khoản `manager` và `tenant`.\n- **manager** chỉ có thể tạo tài khoản `tenant`, và tenant sẽ tự động được gán vào chi nhánh của Manager.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone", "password", "role"],
                "properties": {
                  "phone": { "type": "string", "example": "0912345678", "description": "Số điện thoại (0xxx hoặc +84xxx)" },
                  "password": { "type": "string", "format": "password", "example": "initialPass123" },
                  "full_name": { "type": "string", "example": "Nguyễn Văn A" },
                  "role": { "type": "string", "enum": ["manager", "tenant"], "example": "tenant" },
                  "branch_id": { "type": "integer", "nullable": true, "description": "ID chi nhánh (chỉ super_admin cần điền, manager bị bỏ qua)", "example": 1 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Tạo tài khoản thành công", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" }, "message": { "type": "string" }, "user": { "type": "object", "properties": { "id": { "type": "string", "format": "uuid" }, "phone": { "type": "string" }, "role": { "type": "string" }, "branch_id": { "type": "integer", "nullable": true } } } } } } } },
          "400": { "description": "Thiếu thông tin hoặc số điện thoại đã tồn tại" },
          "401": { "description": "Chưa đăng nhập" },
          "403": { "description": "Không có quyền tạo tài khoản với vai trò này" }
        }
      }
    },
...
    "/api/utility/submit": {
      "post": {
        "tags": ["⚡ Điện nước (Utility)"],
        "summary": "Ghi chỉ số điện nước",
        "description": "Ghi nhận chỉ số điện và nước cho một phòng trong tháng. API sẽ tự động:\n1. So sánh với tháng trước để tính lượng tiêu thụ.\n2. Phát hiện bất thường nếu tăng đột biến >50%.\n\nThường được gọi bởi **Manager** khi đi chốt điện nước.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["roomId", "currentElectricity", "currentWater", "month", "year"],
                "properties": {
                  "roomId": { "type": "integer", "example": 1 },
                  "currentElectricity": { "type": "number", "example": 1523, "description": "Chỉ số điện kế hiện tại (kWh)" },
                  "currentWater": { "type": "number", "example": 45, "description": "Chỉ số nước hiện tại (m³)" },
                  "month": { "type": "integer", "example": 5, "description": "Tháng (1-12)" },
                  "year": { "type": "integer", "example": 2026, "description": "Năm (VD: 2026)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ghi chỉ số thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "log": { "type": "object", "description": "Bản ghi vừa lưu vào database" },
                        "usage": {
                          "type": "object",
                          "properties": {
                            "electricity": { "type": "number", "example": 120, "description": "Số kWh tiêu thụ trong tháng" },
                            "water": { "type": "number", "example": 5, "description": "Số m³ tiêu thụ trong tháng" }
                          }
                        },
                        "anomaly": {
                          "type": "object",
                          "properties": {
                            "detected": { "type": "boolean", "example": false },
                            "messages": { "type": "array", "items": { "type": "string" }, "example": [] }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Chỉ số mới nhỏ hơn tháng trước hoặc thiếu thông tin" },
          "401": { "description": "Chưa đăng nhập" }
        }
      }
    },
...
    "/api/webhooks/payos": {
      "post": {
        "tags": ["💳 Thanh toán (PayOS Webhook)"],
        "summary": "[Chỉ dành cho PayOS] Cập nhật trạng thái thanh toán",
        "description": "⚠️ **API này dành riêng cho PayOS server gọi vào, Mobile App KHÔNG gọi API này.** Khi người dùng thanh toán thành công, PayOS sẽ tự động gọi vào đây để:\n1. Cập nhật trạng thái hóa đơn thành `paid`.\n2. Bắn Push Notification về điện thoại của Tenant.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Payload do PayOS gửi tự động (xem tài liệu PayOS để biết chi tiết)",
                "properties": {
                  "code": { "type": "string", "example": "00" },
                  "orderCode": { "type": "integer", "example": 12345 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Xử lý thành công (luôn trả về 200 để PayOS không gửi lại)" }
        }
      }
    }
  }
}